Two fixes to behaviors (including issue #5) and one test fix:
1.) Should bind to ExecutionResult not a special RequestChildBehavior type,
and do the run on the behavior author's side, not inside the BehaviorBuilder().
This allows the behaviorAuthor to control which feedback and ctx go to the
child behavior, and avoids the weirdness of the builder appearing to use
feedback and ctx twice (which is must happen if it magically supplies feedback
and ctx to the child).

2.) Likewise, inside the behavior body, don't ReturnAction the action you got
back from the binder. We make this mistake more impossible to make by simply
not giving the behavior body the action; all it gets back is Resume (behavior)
or Ready (final result). We might want to rename Ready at some point but for
now it's good to make a clear distinction: it's an input created from an output,
not an output like Finished is.

3.) The justAttack test was placing bobs into the same cell and expecting them
be in range of each other after that, but on a diagonal range was 1.41 and a Move
resulted, which resulted in a flaky test. Placing them exactly 1 yard apart
fixes the flakiness.