wip on animations. Sort of working, but there's a problem--animated nodes don't always go back to their proper position if you go back in time, especially if the animation was still ongoing. (But that's not the only time.) Try 2 ogres vs. 3 orcs and you'll see.
We may want to switch to full animations instead of just tweens,
but I'm not positive that would solve the issue.

I *think* what is happening is that the animated value is left in
place even after we go back in time to set a new Group.x and Group.y.
It's like the animated value takes priority. So how do we end the
animation?