Improve perceived perf by improving responsiveness: more feedback while execution is going on.
Firstly, we have beginFights delay briefly to make sure that even specific fights get a busy signal.

Then, we also make sure that there's a way to inform the user of progress during long fights. Instead of a progress bar, which we can't predict, we just let calibrate() update the busy text.