Okay, part of was going on is that since we always ignore the initial feedback of a behavior, the duplication isn't always hurting us.
The issue was sort of this:

This is an ExecutionResult:
    return feedback

This binds to it and is an 'feedback * 'ctx -> ExecutionResult:
    let! childResult = smokeTest (feedback, ctx)
    return feedback

This binds to that and is a 'ctx -> 'feedback * 'ctx -> ExecutionResult:
    let! ctx = QueryRequest id
    let! childResult = smokeTest (feedback, ctx)
    return feedback

This binds to that and would be 'feedback * 'ctx -> 'ctx * -> 'feedback * 'ctx -> ExecutionResult
    let! feedback, ctx = ReturnAction(feedback)
    let! ctx = QueryRequest id
    let! childResult = smokeTest (feedback, ctx)
    return feedback

You see the problem? functions can be monadic, but this composition method isn't monadic.

So the fix was to create a monadic type that could be all of these things, and the wrinkle
was that trying to use feedback * ctx -> ExecutionResult wasn't appropriate because some
things you bind to (like ReturnAction) don't consume feedback * ctx. And trying to predict
when feedback * ctx would get thrown away was a source of bugs, so we instead make sure that
our monadic type (still called Behavior) can distinguish between cases where we introduce
a requirement for a feedback * ctx and cases where we don't and just need a delay.