Maybe the issue is this:
This is an ExecutionResult:
    return feedback

This binds to it and is an 'feedback * 'ctx -> ExecutionResult:
    let! childResult = smokeTest (feedback, ctx)
    return feedback

This binds to that and is a 'ctx -> 'feedback * 'ctx -> ExecutionResult:
    let! ctx = QueryRequest id
    let! childResult = smokeTest (feedback, ctx)
    return feedback

This binds to that and would be 'feedback * 'ctx -> 'ctx * -> 'feedback * 'ctx -> ExecutionResult
    let! feedback, ctx = ReturnAction(feedback)
    let! ctx = QueryRequest id
    let! childResult = smokeTest (feedback, ctx)
    return feedback

You see the problem? functions can be monadic, but this composition method isn't monadic.

Maybe we should go back and review the monad laws.